RFC: update liteeth driver to automatically calculate hwreg offsets
LiteX MMIO (a.k.a. CSR) registers are exposed in chunks of a given
width (typically 8 or 32), with chunks located at consecutive 32
or 64 bit aligned (depending on the native CPU XLen) locations in
the address space. While alignment can be inferred based on the
hard-coded CONFIG_64BIT kernel option, chunk size must be provided
as a boot-time property (called 'mmio-dw-bytes'), of the ethernet
node's parent (i.e., in our case, the "soc" node).

Make the LiteETH driver work on any combination of CPU XLen and
MMIO chunk size by automatically calculating hardware register
offsets based on alignment and chunk size, the latter of which
defaults to 1 (8 bits) for backward compatibility.

For designs with a wider MMIO/CSR data bus, add the 'mmio-dw-bytes'
property as shown in the following example:

    ...
    L12: soc {
        ...
        mmio-dw-bytes = <4>; /* 32-bit csr-data-width */
        ...
        mac0: mac@92003800 {
            compatible = "litex,liteeth";
            ...
        };
    };
    ...

Signed-off-by: Gabriel Somlo <gsomlo@gmail.com>