emulator: Implement SBI_EXT_BASE syscall.
In trying to boot a v5.8 kernel, it seems that the kernel is attempting
to discover the SBI version using the SBI_EXT_BASE syscall, but this
is resulting in a hung board. I think the kernel is expecting the
syscall to return an error if not supported.

This patch should workaround the issue in two ways: first we implement
the SBI_EXT_BASE syscall to report our version as v0.1, and we also
change the default case to return an error instead of halting the CPU.