ulx3s: use spisdcard instead of sdcard
The ULX3S does not have a card detect (CD) pin for the SD interface,
which breaks the sdcard driver in Linux.

This patch changes the ULX3S to use the "spisdcard" interface instead
of the "sdcard" interface.  With this change the SD card works fine
from the BIOS and Linux in SPI-mode.

Signed-off-by: Curt Brune <curt@brune.net>