make/soc_linux: Change UART baudrate default value to 115200Bauds and add --uart-baudrate parameter.
Now that most of the boards can load Linux images from Ethernet, SDCard or SATA it does
not necessarily make sense to keep the default UART to 1MBauds. UART baudrate now defaults
to 115000Bauds and can be increased manually when it's the only want to load binaries to the
board.