use inferrable power of 2 memory size on Intel boards
For whatever reason, Quartus is unable to infer the integrated_sram
memory when its size is not a power of two. This is confirmed using
Quartus 20.1.1 and Quartus 18.1. As a result, these boards take
excessively long to build and might not fit. The memory size has already
been reduced on some of them in an attempt to fix this issue.

To fix this issue for all of them, the size on all Intel boards is
reduced to the next lowest power of two. I've confirmed that all boards
build, fit, and pass timing using Quartus 20.1.1. I've also confirmed
that De10Nano can still boot the demo image.